Add Animation System to Characteristics section
I started to work on describing the characteristics of the project in
order to provide a better understanding of what makes it special/
important. The first thing that needed to be added was the Animation
system as it is one of the most pominent things of the project.

In order to do this properly, I added the GitHub Resources folder in
order to store the image files that I am going to use moving forward.