Fix problems displaying images on README.md
In adding images to the README, I did not take into consideration the
problems that I would face in displaying them. I changed the img
variable name to src in order to fix bug of images not rendering
properly. I fixed the relative paths of the images in order to
ensure that they can be accessed.