Remove un-used audio (both files and in code)
During the development of my project, I used out various music
tracks and sound effects in order to test out the sound library
that I was using. When I completed the project, some of these
files went unused, but still were present in the resources/music
folder.

I decided to remove them (the files and the mentions in the
code) in order to improve readability and to reduce the overall size
of the project.